requirements.txt
blinker==1.7.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
fonttools==4.51.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
Werkzeug==3.0.2
zipp==3.18.1
